adding a GlobalRunner to avoid repetitive code when running dockers
